Refactor project and adding AudioPlayer
