+++ src/constants/CardType.js
@@ -1,0 +1,5 @@
+export default {
+  LinkAccount: 'LinkAccount',
+  Simple: 'Simple',
+  Standard: 'Standard'
+};
