+++ src/constants/SpeechType.js
@@ -1,0 +1,4 @@
+export default {
+  PlainText: 'PlainText',
+  SSML: 'SSML'
+};
